Added cancelDone, cancelFail, cancelProgress methods + tests. Note that there's no cancelThen yet
The use case for this would be to avoid race conditions when making
serial ajax calls. For example, if i'm using the deferreds returned from
$.ajax to build up a cache, i should be able to cancel all the "on it's
way" callbacks in the cache, instead of having to put a
'exitIfNotLastCall' clause in my callback method.

This turned out relatively easy to implement for done, fail and
progress, since they rely on jQuery callback lists. Might see if we can
do something for .then()

(some discussion from JSBin):
http://jsbin.com/iyelep/5/edit