A way of passing an execution context for a deferred:
var def = jQuery.Deferred();

def.done(this.update, this);
def.fail(this.showError, this);
def.progress(this.progressBar, this);

The inspiration comes from the Backbone.js events mechanisms, which
allows you to pass in execution context. I think this is really handy
and makes the code simpler and easier to read.

This is just a proof of concept to throw the idea out there. I'd be
happy to write tests, etc if people like it